1) Fixed blank table in ARLPlayViewController. 2) Fixed some remaining database context problems. 3) Enlarged Now and satisfiedAt to long long int/longLongValue. 4) Re-enabled saving on MR_defaultContext in background threads (check if this is correct).

+++ ARLearn/Utils/ARLUtils.h
@@ -236,7 +236,7 @@
  *
  *  @return <#return value description#>
  */
-+(long) Now;
++(unsigned long long int) Now;
 
 /*!
  *  Format a double as a Date/Time.
